Refactor main (context) menu into its own class
